test(store): enable USER_RUNTIME_CHECKS token tests (#2602)
